evaluator/requirements.txt
azure-storage-blob
azure-identity
azure-data-tables
azure-functions
azure-storage-queue
azure-cognitiveservices-vision-computervision

gensim==4.3.2
numpy
scipy==1.10.1
